# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "developer@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/user/userinfo": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "取得會員資料",
                "description": "取得會員資料",
                "operationId": "userinfo",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "properties": {
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 12
                                                },
                                                "user_name": {
                                                    "type": "string",
                                                    "example": "paul"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "paul@test.net"
                                                },
                                                "create_time": {
                                                    "type": "integer",
                                                    "example": 1721202502
                                                },
                                                "balance": {
                                                    "type": "integer",
                                                    "example": 200
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error_code": {
                                            "type": "integer",
                                            "example": 401001
                                        },
                                        "error_msg": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "result": false,
                                        "error_code": 401001,
                                        "error_msg": "Unauthenticated."
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/v1/user/payment": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "會員付款",
                "description": "會員付款",
                "operationId": "payment",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "machine_id",
                                    "user_id",
                                    "amount"
                                ],
                                "properties": {
                                    "machine_id": {
                                        "type": "integer",
                                        "example": "12"
                                    },
                                    "amount": {
                                        "type": "integer",
                                        "example": "150"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "properties": {
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 21
                                                },
                                                "machine_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "transaction_amount": {
                                                    "type": "integer",
                                                    "example": 50
                                                },
                                                "after_transaction_balance": {
                                                    "type": "integer",
                                                    "example": 750
                                                },
                                                "transaction_time": {
                                                    "type": "integer",
                                                    "example": 1721540702
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "boolean"
                                        },
                                        "error_code": {
                                            "type": "integer"
                                        },
                                        "error_msg": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "result": false,
                                        "error_code": 400001,
                                        "error_msg": "The JSON payload is not in the correct format."
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/v1/user/newuser": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "新增會員",
                "description": "新會員註冊",
                "operationId": "newuser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_name",
                                    "password",
                                    "email"
                                ],
                                "properties": {
                                    "user_name": {
                                        "type": "string",
                                        "example": "Bob"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "579f2cfcfa83"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "eva.lin321@tems.edu"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "result": true,
                                        "message": "success"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "boolean"
                                        },
                                        "error_code": {
                                            "type": "integer"
                                        },
                                        "error_msg": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "result": false,
                                        "error_code": 400001,
                                        "error_msg": "The JSON payload is not in the correct format."
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}